feat: Update npm dependencies to latest versions
This commit updates the npm dependencies in the frontend/package.json file to their latest stable versions. The updated dependencies include "@babel/eslint-parser" to version "^7.24.7", "chokidar" to version "^3.6.0", "eslint-webpack-plugin" to version "^4.2.0", and "html-webpack-plugin" to version "^5.6.0". This update ensures that the project is using the most up-to-date and secure versions of these dependencies.