refactor: Update userModel.js to include timeStamp and isOnline fields
This commit adds the timeStamp and isOnline fields to the userModel.js file. The timeStamp field is of type Date and is set to the current date and time by default. The isOnline field is of type Boolean and is set to false by default. These additions enhance the user model by providing a timestamp for user creation and a flag to indicate if the user is currently online.