feat: Add background-repeat property to FeaturesSection and Features images
This commit adds the `background-repeat: no-repeat` property to the `FeaturesSection` component and the `Features` images in the project. By setting `background-repeat` to `no-repeat`, we ensure that the background images are not repeated and maintain a consistent visual appearance. This enhancement improves the overall design and user experience of the application.