refactor: Update CORS middleware to allow multiple origins
This commit updates the CORS middleware in the `server.js` file to allow multiple origins. The `allowedOrigins` variable now accepts an array of allowed origins, which can be specified through the `ALLOWED_ORIGINS` environment variable. If the request origin is not in the list of allowed origins, an error is returned.

These changes improve the flexibility and security of the application by allowing requests from multiple origins while still enforcing CORS policies.