refactor: Update feedbackRoutes to remove unnecessary middleware
This commit refactors the feedbackRoutes file to remove the unnecessary middleware functions verifyToken and isPatient. These middleware functions were previously used to authenticate and authorize the user, but they are no longer needed since the authentication and authorization logic has been moved to the createFeedback and getFeedbackForDoctor functions in the feedbackController.

This change simplifies the code and improves the maintainability of the feedbackRoutes file.