refactor: Update DoctorFeedback and FeedbackForm components
This commit updates the DoctorFeedback and FeedbackForm components to improve functionality and usability. In DoctorFeedback.js, the doctorId prop is replaced with the useAuth hook to get the doctor's ID from the userData object. The useEffect dependency is updated to userData._id, and the loading state handling is improved. In FeedbackForm.js, the patientId and doctorId props are replaced with the useAuth hook to get the user's ID from the userData object. The data object is updated to include the userId property instead of patientId and doctorId. These changes enhance the components by using the useAuth hook to retrieve necessary user and doctor information.